compose_creator/requirements.txt
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
PyYAML==6.0.2
